Refactor the UI code
