+++ components/atoms/DropDown/DropDown.tsx
@@ -1,6 +1,7 @@
 import PropTypes, { InferProps } from "prop-types";
 import { ChevronDownIcon, ChevronUpIcon } from "@heroicons/react/24/outline";
 import Image from "next/image";
+import { useState } from "react";
 
 const DropDownProps = {
     classes: PropTypes.string,
@@ -11,7 +12,8 @@ const DropDownProps = {
     options: PropTypes.array
 }
 
-const DropDown = ({ text, classes, isOpen, imageUrl, onClick, options }: InferProps<typeof DropDownProps>) => {
+const DropDown = ({ text, classes, isOpen, imageUrl, onClick, options }: InferProps<typeof DropDownProps>): JSX.Element => {
+    let [makeSelectOption, setMakeSelectOption] = useState<string | null | undefined>(text);
 
     return (
         <>
@@ -21,12 +23,12 @@ const DropDown = ({ text, classes, isOpen, imageUrl, onClick, options }: InferPr
                     {imageUrl && <span className="h-8 w-8 ml-2 -mr-5">
                         <Image src={imageUrl} alt="" />
                     </span>}
-                    <span className="text-sm ml-6">{text}</span>
+                    <span className="text-sm ml-6 capitalize leading-3">{makeSelectOption}</span>
                     <span className="arrow hover:cursor-pointer
                     hover:opacity-100 opacity-80" onClick={onClick}>
                         {!isOpen ? (
                             <ChevronDownIcon className="active:transition active:rotate-180
-                    active:duration-50 active:ease-out h-5 w-5 bold-200  mx-2" />) : (
+                    active:duration-50 active:ease-out h-5 w-5 bold-200 mx-2" />) : (
                             <ChevronUpIcon className="active:transition active:rotate-180
                     active:duration-50 active:ease-out  h-5 w-5 bold-200 mx-2" />)
                         }
@@ -39,7 +41,8 @@ const DropDown = ({ text, classes, isOpen, imageUrl, onClick, options }: InferPr
                      bg-slate-200 px-4 py-1 w-[100%]  my-0 rounded-sm">
                         {options?.map((optText) => (
                             <span key={optText} className="text-sm block capitalize w-100 my-4
-                             hover:cursor-pointer hover:underline"> {optText} </span>
+                             hover:cursor-pointer hover:underline"
+                                onClick={() => setMakeSelectOption(optText)}> {optText} </span>
                         ))}
                     </div>
                 )}
@@ -49,3 +52,7 @@ const DropDown = ({ text, classes, isOpen, imageUrl, onClick, options }: InferPr
 }
 
 export default DropDown
+
+function setMakeSelectOption(e: any) {
+    throw new Error("Function not implemented.");
+}
